- fixed some bugs

+++ src/main/java/com/spadsystem/Home.java
@@ -6,24 +6,32 @@ import com.spadsystem.model.User;
 import org.glassfish.jersey.media.multipart.*;
 import org.json.JSONObject;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import java.io.*;
 import java.net.URI;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
 
 @Path("/")
 public class Home {
     private String dir = this.getClass().getProtectionDomain().getCodeSource().getLocation().getPath();
 
+    @Context
+    HttpServletRequest httpServletRequest;
+
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String test() {
-        JDBC.fromExcel();
+//        JDBC.fromExcel();
 //        JDBC.clearDB();
         return "server is GOOD!";
     }
@@ -48,12 +56,17 @@ public class Home {
         User user = new User();
         user.setUser_id(id);
         user.setPassword(password);
+        HttpSession session = httpServletRequest.getSession(true);
         try {
             if (JDBC.getInstance().checkUser(user)) {
+                session.setAttribute("spadsystem-user", id);
+                session.setAttribute("spadsystem-time", new Date());
                 if (user.getType() != null && user.getType().equals("true"))
-                    return Response.seeOther(URI.create("/spadsystem/admin.html")).build();
+                    session.setAttribute("spadsystem-type", "admin");
                 else
-                    return Response.seeOther(URI.create("/spadsystem/index.jsp?id=" + id + "&type=" + user.getType())).build();
+                    session.setAttribute("spadsystem-type", "user");
+                return Response.seeOther(URI.create("/spadsystem/index.jsp?id=" + id + "&type=" + user.getType())).build();
+
             } else
                 return Response.seeOther(URI.create("/spadsystem/login.html?text=auth")).build();
         } catch (ClassNotFoundException | SQLException e) {
@@ -66,8 +79,12 @@ public class Home {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public List<String> firstData() {
+        String id = httpServletRequest.getQueryString().substring(3);
         try {
-            return JDBC.getInstance().getFirstData();
+            HttpSession session = httpServletRequest.getSession(true);
+            if (session.getAttribute("spadsystem-user") != null &&
+                    session.getAttribute("spadsystem-user").equals(id))
+                return JDBC.getInstance().getFirstData();
         } catch (ClassNotFoundException | SQLException e) {
             e.printStackTrace();
         }
@@ -78,9 +95,12 @@ public class Home {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.TEXT_PLAIN)
-    public boolean addUser(User user) {
+    public boolean addUser(User user) throws Throwable {
         try {
-            return JDBC.getInstance().addUser(user);
+            HttpSession session = httpServletRequest.getSession(true);
+            if (session.getAttribute("spadsystem-user") != null &&
+                    session.getAttribute("spadsystem-type").equals("admin"))
+                return JDBC.getInstance().addUser(user);
         } catch (SQLException | ClassNotFoundException e) {
             e.printStackTrace();
         }
@@ -91,9 +111,12 @@ public class Home {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.TEXT_PLAIN)
-    public User updateUserPassword(User user) {
+    public User updateUserPassword(User user) throws Throwable {
         try {
-            return JDBC.getInstance().updateUserPassword(user);
+            HttpSession session = httpServletRequest.getSession(true);
+            if (session.getAttribute("spadsystem-user") != null &&
+                    session.getAttribute("spadsystem-user").equals(user.getUser_id()))
+                return JDBC.getInstance().updateUserPassword(user);
         } catch (SQLException | ClassNotFoundException e) {
             e.printStackTrace();
         }
@@ -106,7 +129,10 @@ public class Home {
     @Produces(MediaType.TEXT_PLAIN)
     public boolean deleteUser(User user) {
         try {
-            return JDBC.getInstance().deleteUser(user);
+            HttpSession session = httpServletRequest.getSession(true);
+            if (session.getAttribute("spadsystem-user") != null &&
+                    session.getAttribute("spadsystem-type").equals("admin"))
+                return JDBC.getInstance().deleteUser(user);
         } catch (SQLException | ClassNotFoundException e) {
             e.printStackTrace();
         }
@@ -120,21 +146,24 @@ public class Home {
     public JSONObject uploadImage(@FormDataParam("file") InputStream uploadedInputStream,
                                   @FormDataParam("file") FormDataContentDisposition fileDetails,
                                   @FormDataParam("id") String id) {
-        String uploadedFileLocation = dir.substring(0, dir.indexOf("WEB-INF")) + "temp/img_" + id + ".png";
-        // save it
-        try {
-            OutputStream out = new FileOutputStream(new File(uploadedFileLocation));
-            int read;
-            byte[] bytes = new byte[1024];
-            while ((read = uploadedInputStream.read(bytes)) != -1) {
-                out.write(bytes, 0, read);
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null) {
+            String uploadedFileLocation = dir.substring(0, dir.indexOf("WEB-INF")) + "temp/img_" + id + ".png";
+            // save it
+            try {
+                OutputStream out = new FileOutputStream(new File(uploadedFileLocation));
+                int read;
+                byte[] bytes = new byte[1024];
+                while ((read = uploadedInputStream.read(bytes)) != -1) {
+                    out.write(bytes, 0, read);
+                }
+                JDBC.getInstance().saveImage(id, uploadedFileLocation);
+                out.flush();
+                out.close();
+                return new JSONObject("{\"File uploaded to\" : \"" + uploadedFileLocation + "\"}");
+            } catch (IOException | SQLException | ClassNotFoundException e) {
+                e.printStackTrace();
             }
-            JDBC.getInstance().saveImage(id, uploadedFileLocation);
-            out.flush();
-            out.close();
-            return new JSONObject("{\"File uploaded to\" : \"" + uploadedFileLocation + "\"}");
-        } catch (IOException | SQLException | ClassNotFoundException e) {
-            e.printStackTrace();
         }
         return new JSONObject("detail:Error");
     }
@@ -144,6 +173,14 @@ public class Home {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public List<User> search(String data) {
+        if (data.contains("all")) {
+            HttpSession session = httpServletRequest.getSession(true);
+            if (session.getAttribute("spadsystem-user") == null)
+                return null;
+            if (!session.getAttribute("spadsystem-type").equals("admin")) {
+                return null;
+            }
+        }
         try {
             return JDBC.getInstance().searchDb(new JSONObject(data).getString("data").split(" "));
         } catch (ClassNotFoundException | SQLException e) {
@@ -156,13 +193,16 @@ public class Home {
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public User getUser(String string) {
-        User user = new User();
-        user.setUser_id(new JSONObject(string).getString("user_id"));
-        try {
-            return JDBC.getInstance().getUser(user);
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
+    public User getUser() {
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null) {
+            User user = new User();
+            user.setUser_id((String) session.getAttribute("spadsystem-user"));
+            try {
+                return JDBC.getInstance().getUser(user);
+            } catch (ClassNotFoundException | SQLException e) {
+                e.printStackTrace();
+            }
         }
         return null;
     }
@@ -187,11 +227,14 @@ public class Home {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public Mail sendMail(Mail mail) {
-        try {
-            return JDBC.getInstance().saveMail(mail);
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
-        }
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null &&
+                session.getAttribute("spadsystem-user").equals(mail.getSender().getUser_id()))
+            try {
+                return JDBC.getInstance().saveMail(mail);
+            } catch (ClassNotFoundException | SQLException e) {
+                e.printStackTrace();
+            }
         return null;
     }
 
@@ -200,24 +243,27 @@ public class Home {
     @Produces(MediaType.TEXT_PLAIN)
     @Consumes(MediaType.MULTIPART_FORM_DATA)
     public String sendImage(@FormDataParam("file") FormDataBodyPart files) {
-        String obj = this.getClass().getProtectionDomain().getCodeSource().getLocation().getPath();
-        OutputStream out;
-        int read;
-        byte[] bytes = new byte[1024];
-        for (BodyPart part : files.getParent().getBodyParts()) {
-            InputStream is = part.getEntityAs(InputStream.class);
-            ContentDisposition meta = part.getContentDisposition();
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null) {
+            String obj = this.getClass().getProtectionDomain().getCodeSource().getLocation().getPath();
+            OutputStream out;
+            int read;
+            byte[] bytes = new byte[1024];
+            for (BodyPart part : files.getParent().getBodyParts()) {
+                InputStream is = part.getEntityAs(InputStream.class);
+                ContentDisposition meta = part.getContentDisposition();
 
-            String uploadedFileLocation = obj.substring(0, obj.indexOf("WEB-INF")) + "temp/" + meta.getFileName();
-            try {
-                out = new FileOutputStream(new File(uploadedFileLocation));
-                while ((read = is.read(bytes)) != -1) {
-                    out.write(bytes, 0, read);
+                String uploadedFileLocation = obj.substring(0, obj.indexOf("WEB-INF")) + "temp/" + meta.getFileName();
+                try {
+                    out = new FileOutputStream(new File(uploadedFileLocation));
+                    while ((read = is.read(bytes)) != -1) {
+                        out.write(bytes, 0, read);
+                    }
+                    out.flush();
+                    out.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
-                out.flush();
-                out.close();
-            } catch (IOException e) {
-                e.printStackTrace();
             }
         }
         return "image saved";
@@ -228,11 +274,14 @@ public class Home {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public List<Mail> getMail(Mail mail) {
-        try {
-            return JDBC.getInstance().getMail(mail.getReceiver().getUser_id());
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
-        }
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null &&
+                session.getAttribute("spadsystem-user").equals(mail.getReceiver().getUser_id()))
+            try {
+                return JDBC.getInstance().getMail(mail.getReceiver().getUser_id());
+            } catch (ClassNotFoundException | SQLException e) {
+                e.printStackTrace();
+            }
         return null;
     }
 
@@ -242,11 +291,13 @@ public class Home {
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public Mail updateMail(Mail mail) {
-        try {
-            return JDBC.getInstance().updateMail(mail);
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
-        }
+        HttpSession session = httpServletRequest.getSession(true);
+        if (session.getAttribute("spadsystem-user") != null)
+            try {
+                return JDBC.getInstance().updateMail(mail);
+            } catch (ClassNotFoundException | SQLException e) {
+                e.printStackTrace();
+            }
         return null;
     }
 }
